--- assets/styles/_wcs.scss
@@ -1,40 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Varela+Round&display=swap');
-$light: #ff6c71;
-
-* {
-    box-sizing: border-box;
-}
-
-body {
-    margin: 0;
-    background-color: $wild-color;
-    color: $light;
-}
-
-
-.pink-stripe {
-    height: 50vh;
-    width: 100%;
-    display: flex;
-    font-family: 'Varela Round', sans-serif;
-    text-align: center;
-    font-size: clamp(2rem, 5vw, 3rem);
-    align-items: center;
-    justify-content: center;
-}
-
-.white-stripe {
-    background-color: white;
-    padding: 1rem;
-    width: 100%;
-    height: 50vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    img {
-        width: 500px;
-        max-width: 100%;
-        height: auto;
-    }
-}
