+++ templates/base.html.twig
@@ -7,7 +7,6 @@
 			{% block title %}Externatic
 			{% endblock %}
 		</title>
-		{# Run `composer require symfony/webpack-encore-bundle` to start using Symfony UX #}
 		{% block stylesheets %}
 			{{ encore_entry_link_tags('app') }}
 		{% endblock %}
